Create component for heading
